Update s3backup to use a single command and initialize itself
In order to properly run the `s3backup` command so that it can be ran as
a single command this commit checks to see if the
`S3_BACKUP_SINGLE_COMMAND` environment variable is set to enable it.
This commit also adds an `init` command to initialize the configuration
for the `s3backup` command.  The readme was also updated with *Usage*
documentation in this commit.  The version was bumped to `2.0.0` due to
all the added changes that will make it incompatible with previous
versions.